ajout des droits d'accès pour la fonction deleteCustomer dans le controller customers.js

+++ Controllers/customers.js
@@ -1,4 +1,46 @@
 const Customer = require('../Models/Customer');
+const Advisor = require('../Models/Advisor');
+const Manager = require('../Models/Manager');
+const Director = require('../Models/Director');
+const mongoose = require('mongoose');
+
+
+const checkAdvisorId = (advisor_id, customer_id) => {
+    return Customer.findOne({ advisorId: advisor_id, _id: customer_id })
+    .then((customer) => {
+        if(customer) return true;
+        else return false;
+    })
+}
+const checkManagerId = async (manager_id, customer) => {
+    let result = 0;
+    let validate = false;
+    let customersFound = [];
+    let advisorsFound = await Advisor.find({ managerId: manager_id });
+  
+    for (let i = 0; i < advisorsFound.length; i++) {
+        const element = advisorsFound[i];
+        let customer_found = await Customer.find({ advisorId: element._id });
+        customersFound.push(customer_found);
+    }
+
+    customersFound.forEach(element => {
+        element.forEach(el => {
+            if (customer._id.toString() === el._id.toString()) {
+                result +=1;
+                if (result === 1) validate = true;
+            }
+        })
+    })
+    return validate;
+}
+const checkDirectorId = async (director_id) => {
+    return Director.find({ _id: director_id })
+    .then((director) => {
+        if(director) return true;
+        else return false;
+    })
+}
 
 const createCustomer = (req, res) => {
     const newCustomer = new Customer({
@@ -20,17 +62,18 @@ const createCustomer = (req, res) => {
 }
 
 const getCustomers = (req, res) => {
-    /*let criteria = {};
-    const city = req.query.adress.city;
-    if (city === undefined) {
+    let criteria = {};
+    // const city = req.query.adress;
+    /*if (city === undefined) {
         criteria = {};
         console.log(criteria.city+"-----");
     } else {
         console.log(criteria.city+"+++");
         criteria.city = city ;
     }*/
-    return Customer.find()
+    return Customer.find(criteria)
     .then((customers) => {
+        // console.log(customers[3].adress.city);
         return res.send(customers);
     })
     .catch((error) => {
@@ -72,17 +115,56 @@ const putCustomer = (req, res) => {
     })
 }
 
-const deleteCustomer = (req, res) => {
+const deleteCustomer = async (req, res) => {
     const id = req.params.id;
-    return Customer.findByIdAndDelete(id)
-    .then((customer) => {
-        return res.send(`Customer ${customer.firstname} ${customer.lastname} has been deleted.`);
-    })
-    .catch((error) => {
-        return res.status(400).send(error);
-    })
+    const userId = req.headers.userid;
+    const isValidUser = mongoose.isValidObjectId(userId);
+    const isValidCustomer = mongoose.isValidObjectId(id);
+    if (isValidCustomer === false) {
+        return res.status(400).send("l'id du customer n'est pas valide");
+    }
+    const customer = await Customer.findById(id);
+    if (isValidCustomer === true) {
+        const advisor = await Advisor.findById(userId);
+        const manager = await Manager.findById(userId);
+        const director = await Director.findById(userId);
+        let checkAdvisor;
+        let checkManager;
+        let checkDirector;
+        if (advisor != null || advisor != undefined) {
+            try {
+                checkAdvisor = await checkAdvisorId(advisor._id, customer._id);
+            } catch (error) {
+                return res.status(500).send(error);
+            }
+        }
+        if (manager != null || manager != undefined) {
+            try{
+                checkManager = await checkManagerId(manager._id, customer);
+            } catch (error) {
+                return res.status(500).send(error);
+            }
+        }
+        if (director != null || director != undefined) {
+            try {
+                checkDirector = await checkDirectorId(director._id);
+            } catch (error) {
+                return res.status(500).send(error);
+            }
+        }
+        if (checkAdvisor === true || checkManager === true || checkDirector === true && isValidUser === true) {
+            return Customer.findByIdAndDelete(id)
+            .then(() => {
+                return res.send(`Customer ${customer.firstname} ${customer.lastname} has been deleted.`);
+            })
+            .catch((error) => {
+                return res.status(400).send(error);
+            })
+        } else {
+            return res.status(403).send('You don\'t have access to this customer\'s data !');
+        }
+    }
 }
-
 module.exports = {
     createCustomer: createCustomer,
     getCustomers: getCustomers,
